feat: removed smelly code
[IN-543]
